requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
click==8.1.7
python-dateutil==2.8.2
psycopg[binary,pool]==3.1.18
python-dotenv==1.0.0
